+++ driven/storage/adapter.go
@@ -3,6 +3,7 @@ package storage
 import (
 	"context"
 	"core-building-block/core/model"
+	"fmt"
 	"strconv"
 	"sync"
 	"time"
@@ -29,6 +30,12 @@ type Adapter struct {
 
 	cachedApplications *syncmap.Map
 	applicationsLock   *sync.RWMutex
+
+	cachedAuthTypes *syncmap.Map
+	authTypesLock   *sync.RWMutex
+
+	cachedApplicationsOrganizations *syncmap.Map //cache applications organizations
+	applicationsOrganizationsLock   *sync.RWMutex
 }
 
 //Start starts the storage
@@ -53,6 +60,18 @@ func (sa *Adapter) Start() error {
 	err = sa.cacheApplications()
 	if err != nil {
 		return errors.WrapErrorAction(logutils.ActionCache, model.TypeApplication, nil, err)
+	}
+
+	//cache the auth types
+	err = sa.cacheAuthTypes()
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionCache, model.TypeAuthType, nil, err)
+	}
+
+	//cache the application organization
+	err = sa.cacheApplicationsOrganizations()
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionCache, model.TypeApplicationOrganization, nil, err)
 	}
 
 	return err
@@ -215,6 +234,107 @@ func (sa *Adapter) getCachedApplicationTypeByIdentifier(appTypeIdentifier string
 	return nil, nil, errors.ErrorData(logutils.StatusMissing, model.TypeApplicationType, &logutils.FieldArgs{"identifier": appTypeIdentifier})
 }
 
+//cacheAuthTypes caches the auth types
+func (sa *Adapter) cacheAuthTypes() error {
+	sa.logger.Info("cacheAuthTypes..")
+
+	authTypes, err := sa.LoadAuthTypes()
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionFind, model.TypeAuthType, nil, err)
+	}
+
+	sa.setCachedAuthTypes(authTypes)
+
+	return nil
+}
+
+func (sa *Adapter) setCachedAuthTypes(authProviders []model.AuthType) {
+	sa.authTypesLock.Lock()
+	defer sa.authTypesLock.Unlock()
+
+	sa.cachedAuthTypes = &syncmap.Map{}
+	validate := validator.New()
+
+	for _, authType := range authProviders {
+		err := validate.Struct(authType)
+		if err == nil {
+			//we will get it by id and code as well
+			sa.cachedAuthTypes.Store(authType.ID, authType)
+			sa.cachedAuthTypes.Store(authType.Code, authType)
+		} else {
+			sa.logger.Errorf("failed to validate and cache auth type with code %s: %s", authType.Code, err.Error())
+		}
+	}
+}
+
+func (sa *Adapter) getCachedAuthType(key string) (*model.AuthType, error) {
+	sa.authTypesLock.RLock()
+	defer sa.authTypesLock.RUnlock()
+
+	errArgs := &logutils.FieldArgs{"code or id": key}
+
+	item, _ := sa.cachedAuthTypes.Load(key)
+	if item != nil {
+		authType, ok := item.(model.AuthType)
+		if !ok {
+			return nil, errors.ErrorAction(logutils.ActionCast, model.TypeAuthType, errArgs)
+		}
+		return &authType, nil
+	}
+	return nil, errors.ErrorData(logutils.StatusMissing, model.TypeOrganization, errArgs)
+}
+
+//cacheApplicationsOrganizations caches the applications organizations
+func (sa *Adapter) cacheApplicationsOrganizations() error {
+	sa.logger.Info("cacheApplicationsOrganizations..")
+
+	applicationsOrganizations, err := sa.LoadApplicationsOrganizations()
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionFind, model.TypeApplicationOrganization, nil, err)
+	}
+
+	sa.setCachedApplicationsOrganizations(applicationsOrganizations)
+
+	return nil
+}
+
+func (sa *Adapter) setCachedApplicationsOrganizations(applicationsOrganization []model.ApplicationOrganization) {
+	sa.applicationsOrganizationsLock.Lock()
+	defer sa.applicationsOrganizationsLock.Unlock()
+
+	sa.cachedApplicationsOrganizations = &syncmap.Map{}
+	validate := validator.New()
+
+	for _, appOrg := range applicationsOrganization {
+		err := validate.Struct(appOrg)
+		if err == nil {
+			key := fmt.Sprintf("%s_%s", appOrg.Application.ID, appOrg.Organization.ID)
+			sa.cachedApplicationsOrganizations.Store(key, appOrg)
+		} else {
+			sa.logger.Errorf("failed to validate and cache applications organizations with ids %s-%s: %s",
+				appOrg.Application.ID, appOrg.Organization.ID, err.Error())
+		}
+	}
+}
+
+func (sa *Adapter) getCachedApplicationOrganization(appID string, orgID string) (*model.ApplicationOrganization, error) {
+	sa.applicationsOrganizationsLock.RLock()
+	defer sa.applicationsOrganizationsLock.RUnlock()
+
+	key := fmt.Sprintf("%s_%s", appID, orgID)
+	errArgs := &logutils.FieldArgs{"key": key}
+
+	item, _ := sa.cachedApplicationsOrganizations.Load(key)
+	if item != nil {
+		appOrg, ok := item.(model.ApplicationOrganization)
+		if !ok {
+			return nil, errors.ErrorAction(logutils.ActionCast, model.TypeApplicationOrganization, errArgs)
+		}
+		return &appOrg, nil
+	}
+	return nil, errors.ErrorData(logutils.StatusMissing, model.TypeApplicationOrganization, errArgs)
+}
+
 //LoadAuthTypes loads all auth types
 func (sa *Adapter) LoadAuthTypes() ([]model.AuthType, error) {
 	filter := bson.D{}
@@ -228,6 +348,89 @@ func (sa *Adapter) LoadAuthTypes() ([]model.AuthType, error) {
 	}
 
 	return result, nil
+}
+
+//FindAuthType finds auth type by id or code
+func (sa *Adapter) FindAuthType(codeOrID string) (*model.AuthType, error) {
+	return sa.getCachedAuthType(codeOrID)
+}
+
+//InsertLoginSession inserts login session
+func (sa *Adapter) InsertLoginSession(loginSession model.LoginSession) (*model.LoginSession, error) {
+	storageLoginSession := loginSessionToStorage(loginSession)
+
+	_, err := sa.db.loginsSessions.InsertOne(storageLoginSession)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionInsert, model.TypeLoginSession, nil, err)
+	}
+
+	return &loginSession, nil
+}
+
+//FindLoginSession finds login session by refresh token
+func (sa *Adapter) FindLoginSession(refreshToken string) (*model.LoginSession, error) {
+	//find loggin session
+	filter := bson.D{primitive.E{Key: "refresh_token", Value: refreshToken}}
+	var loginsSessions []loginSession
+	err := sa.db.loginsSessions.Find(filter, &loginsSessions, nil)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeLoginSession, nil, err)
+	}
+	if len(loginsSessions) == 0 {
+		//not found
+		return nil, nil
+	}
+	loginSession := loginsSessions[0]
+
+	//account - from storage
+	var account *model.Account
+	if loginSession.AccountAuthTypeID != nil {
+		account, err = sa.FindAccountByID(loginSession.Identifier)
+		if err != nil {
+			return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAccount, nil, err)
+		}
+	}
+
+	//auth type - from cache
+	authType, err := sa.getCachedAuthType(loginSession.AuthTypeCode)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAuthType, nil, err)
+	}
+
+	//application organization - from cache
+	appOrg, err := sa.getCachedApplicationOrganization(loginSession.AppID, loginSession.OrgID)
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeApplicationOrganization, nil, err)
+	}
+
+	modelLoginSession := loginSessionFromStorage(loginSession, *authType, account, *appOrg)
+	return &modelLoginSession, nil
+}
+
+//UpdateLoginSession updates login session
+func (sa *Adapter) UpdateLoginSession(loginSession model.LoginSession) error {
+	storageLoganSession := loginSessionToStorage(loginSession)
+
+	filter := bson.D{primitive.E{Key: "_id", Value: storageLoganSession.ID}}
+	err := sa.db.loginsSessions.ReplaceOne(filter, storageLoganSession, nil)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionUpdate, model.TypeLoginSession, &logutils.FieldArgs{"_id": storageLoganSession.ID}, err)
+	}
+
+	return nil
+}
+
+//DeleteLoginSession deletes login session
+func (sa *Adapter) DeleteLoginSession(id string) error {
+	filter := bson.M{"_id": id}
+	res, err := sa.db.loginsSessions.DeleteOne(filter, nil)
+	if err != nil {
+		return errors.WrapErrorAction(logutils.ActionDelete, model.TypeLoginSession, nil, err)
+	}
+	if res.DeletedCount != 1 {
+		return errors.ErrorAction(logutils.ActionDelete, model.TypeLoginSession, logutils.StringArgs("unexpected deleted count"))
+	}
+	return nil
 }
 
 //FindAccount finds an account for app, org, auth type and account auth type identifier
@@ -726,7 +929,7 @@ func (sa *Adapter) FindRefreshToken(token string) (*model.AuthRefresh, error) {
 	filter := bson.M{"$or": conditions}
 
 	var refresh model.AuthRefresh
-	err := sa.db.refreshTokens.FindOne(filter, &refresh, nil)
+	err := sa.db.loginsSessions.FindOne(filter, &refresh, nil)
 	if err != nil {
 		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAuthRefresh, nil, err)
 	}
@@ -740,7 +943,7 @@ func (sa *Adapter) LoadRefreshTokens(orgID string, appID string, credsID string)
 	opts := options.Find().SetSort(bson.D{primitive.E{Key: "exp", Value: 1}})
 
 	var refresh []model.AuthRefresh
-	err := sa.db.refreshTokens.Find(filter, &refresh, opts)
+	err := sa.db.loginsSessions.Find(filter, &refresh, opts)
 	if err != nil {
 		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAuthRefresh, nil, err)
 	}
@@ -754,7 +957,7 @@ func (sa *Adapter) InsertRefreshToken(refresh *model.AuthRefresh) error {
 		return errors.ErrorData(logutils.StatusInvalid, model.TypeAuthRefresh, nil)
 	}
 
-	_, err := sa.db.refreshTokens.InsertOne(refresh)
+	_, err := sa.db.loginsSessions.InsertOne(refresh)
 	if err != nil {
 		return errors.WrapErrorAction(logutils.ActionInsert, model.TypeAuthRefresh, nil, err)
 	}
@@ -775,7 +978,7 @@ func (sa *Adapter) UpdateRefreshToken(token string, refresh *model.AuthRefresh) 
 		}},
 	}
 
-	res, err := sa.db.refreshTokens.UpdateOne(filter, update, nil)
+	res, err := sa.db.loginsSessions.UpdateOne(filter, update, nil)
 	if err != nil {
 		return errors.WrapErrorAction(logutils.ActionFind, model.TypeAuthRefresh, nil, err)
 	}
@@ -790,7 +993,7 @@ func (sa *Adapter) UpdateRefreshToken(token string, refresh *model.AuthRefresh) 
 func (sa *Adapter) DeleteRefreshToken(token string) error {
 	filter := bson.D{primitive.E{Key: "current_token", Value: token}}
 
-	res, err := sa.db.refreshTokens.DeleteOne(filter, nil)
+	res, err := sa.db.loginsSessions.DeleteOne(filter, nil)
 	if err != nil {
 		return errors.WrapErrorAction(logutils.ActionDelete, model.TypeAuthRefresh, nil, err)
 	}
@@ -805,7 +1008,7 @@ func (sa *Adapter) DeleteRefreshToken(token string) error {
 func (sa *Adapter) DeleteExpiredRefreshTokens(now *time.Time) error {
 	filter := bson.M{"exp": bson.M{"$lte": now}}
 
-	_, err := sa.db.refreshTokens.DeleteMany(filter, nil)
+	_, err := sa.db.loginsSessions.DeleteMany(filter, nil)
 	if err != nil {
 		return errors.WrapErrorAction(logutils.ActionDelete, model.TypeAuthRefresh, &logutils.FieldArgs{"exp": now}, err)
 	}
@@ -1374,6 +1577,11 @@ func (sa *Adapter) LoadApplicationsOrganizations() ([]model.ApplicationOrganizat
 	return result, nil
 }
 
+//FindApplicationOrganizations finds application organization
+func (sa *Adapter) FindApplicationOrganizations(appID string, orgID string) (*model.ApplicationOrganization, error) {
+	return sa.getCachedApplicationOrganization(appID, orgID)
+}
+
 // ============================== ServiceRegs ==============================
 
 //FindServiceRegs fetches the requested service registration records
@@ -1553,14 +1761,26 @@ func NewStorageAdapter(mongoDBAuth string, mongoDBName string, mongoTimeout stri
 	cachedApplications := &syncmap.Map{}
 	applicationsLock := &sync.RWMutex{}
 
+	cachedAuthTypes := &syncmap.Map{}
+	authTypesLock := &sync.RWMutex{}
+
+	cachedApplicationsOrganizations := &syncmap.Map{}
+	applicationsOrganizationsLock := &sync.RWMutex{}
+
 	db := &database{mongoDBAuth: mongoDBAuth, mongoDBName: mongoDBName, mongoTimeout: timeout, logger: logger}
 	return &Adapter{db: db, logger: logger, cachedOrganizations: cachedOrganizations, organizationsLock: organizationsLock,
-		cachedApplications: cachedApplications, applicationsLock: applicationsLock}
+		cachedApplications: cachedApplications, applicationsLock: applicationsLock,
+		cachedAuthTypes: cachedAuthTypes, authTypesLock: authTypesLock,
+		cachedApplicationsOrganizations: cachedApplicationsOrganizations, applicationsOrganizationsLock: applicationsOrganizationsLock}
 }
 
 type storageListener struct {
 	adapter *Adapter
 	DefaultListenerImpl
+}
+
+func (sl *storageListener) OnAuthTypesUpdated() {
+	sl.adapter.cacheAuthTypes()
 }
 
 func (sl *storageListener) OnOrganizationsUpdated() {
@@ -1575,6 +1795,7 @@ func (sl *storageListener) OnApplicationsUpdated() {
 func (sl *storageListener) OnApplicationsOrganizationsUpdated() {
 	sl.adapter.cacheApplications()
 	sl.adapter.cacheOrganizations()
+	sl.adapter.cacheApplicationsOrganizations()
 }
 
 //Listener represents storage listener